python-agent-basic/requirements.txt
xenarch[agent]
python-dotenv
httpx
